// frontend/src/contracts/types.ts
export interface ApiResponse<T> {
  code: string
  message: string
  data: T
  request_id: string
}

export interface ApiFieldError {
  field: string
  reason: string
}